mpsc-queue: remove spinlock implementation
A dumb userland spinlock implementation is worthless and only
consumes bytes from the queue.

It would only be useful if it was attached to a type that could
be used to enforce thread safety by the compiler, making sure
that the API is respected.

Simplify and remove instead of further complicating the queue.

Signed-off-by: Gaetan Rivet <grive@u256.net>